requirements.txt
matplotlib==3.9.2
numpy==2.1.2
Pillow==11.0.0
scikit_learn==1.5.2
ucimlrepo==0.0.7
opencv-python==4.10.0
